fix: #2255 (#2446)

+++ Intersect.Client.Core/MonoGame/Network/MonoSocket.cs
@@ -251,7 +251,7 @@ internal partial class MonoSocket : GameSocket
                     );
                 }
 
-                if (MainMenu.LastNetworkStatusChangeTime + ServerStatusPingInterval * 1.5f < now)
+                if (MainMenu.LastNetworkStatusChangeTime + (int)(ServerStatusPingInterval * 1.5f) < now)
                 {
                     MainMenu.SetNetworkStatus(NetworkStatus.Offline);
                 }
